renderer/multigpu: Don't unwrap renderer error in `import_shm_buffer`
This can error due to the buffer type being something other than an shm
buffer, but it can also error for other reasons, like
`BufferAccessError::BadMap`.

This fixes a panic when a client tries to use a buffer wit a file that
isn't large enough. With this, there's no longer a panic and the client
gets a protocol error as it is supposed to.

Fixes https://github.com/Smithay/smithay/issues/834.