Apply protocol state in post-commit hooks
State should be applied on the actual commit. Pre-commit is too early,
because the actual commit can and will be delayed by blockers. This
would lead to inconsistent state where the current surface state is
still old, but the protocol surface state is already new. Post-commit
hook is the right place for this, it runs right after the current
surface state is updated, and before the compositor commit handler is
called.