drm/atomic: Put `RwLock<PropMapping>` in `Arc`, and share
If this has to be in an `RwLock` anyway, it seems unnecessary to clone
it, and it can just be shared.

(`AtomicDrmDevice` does not use `RwLock` yet, but will if it wants to
update the mapping.)

I don't expect this to cause problems, though maybe it doesn't help much
either. But it makes sense.