keyboard: Combine xkb types in one struct
It's cleaner to make the `unsafe impl Send` more fine-grained, and this
should help with some future changes.

Since `XkbContextHandler` is public, this fixes a soundness issue in the
public API, since it was possible with a safe call to get an
`xkb::Keymap` and `xkb::State`, and clone those. Potentially violating
the safety assumption of the `Send` implementation.