drm-syncobj: Fix handling of `wp_linux_drm_syncobj_surface_v1::destroy`
It should be possible to destroy this, and then create a new instance
for the same `wl_surface`. And it seems that does happen with Firefox on
Nvidia.

I guess I meant to add some handling to `Destroy`, but didn't do so
earlier. This removes the hooks, signals any pending release fence,
and allows `get_surface` to be called again.

This should be consistent with how the protocol spec says destroy should
behave. If a `wl_surface::commit` has already occurred, the sync points
should still be used when presenting that commit.