WIP Enable `clippy::same_name_method` lint
It is confusing and error prone to have a method on a type and a trait
implemented by that type with the same name.

This is particularly bad for something like `<Window as SpaceElement>::bbox`,
which actually calls `Window::bbox_with_popups`. Which is different from
`Window::box`. But these three functions have the same type signature.

It seems good to disallow this lint in general, but it can be allowed
where the methods do the same thing.