gles: Define a `TextureShaderElement`
`GlesFrame` has support for using a custom shader, but there doesn't
seem to be a particularly convenient way to use this. I'm not sure
exactly how `override_default_tex_program` would best be used, and it's
probably not a good way to use a shader for one specific element.

`TextureShaderElement` wraps a `TextureRenderElement` with a shader and
uniforms. So doesn't have to duplicate the implementation of different
constructors `TextureRenderElement` provides.