input: Make `TabletToolEvent` depend on `AbsolutePositionEvent`
`TabletToolEvent` provides all the methods that are provide by
`AbsolutePositionEvent`. Seemingly with the only difference that tablet
events guarantee the untransformed coordinates are in mm, while it is
otherwise unspecified with `AbsolutePositionEvent`.

Re-using the `AbsolutePositionEvent` trait here makes it easier to
share a bit of code for handling tablet and touch events.