pixman: Store strong reference to `WlBuffer` in `PixmanImage`
If a weak `WlBuffer` is used, it isn't possible to read the contents of
a destroyed shm buffer. But the Wayland protocol specifies that
destroying an attached buffer shouldn't change the contents of the
surface.

A `wayland_server::Weak` will not upgrade to a strong reference for a
destroyed object even if other strong references exist, like the strong
reference to the `WlBuffer` in `WaylandSurfaceRenderElement`.

This fixes an error when a buffer is destroyed by libdecor, which
crashed a version of Anvil patched to use pixman.