utils/geometry: Deprecate `::from_loc_and_size` in favor of a `::new`
This new function is the same as `from_loc_and_size`, except it takes
concrete types as arguments instead of `impl Into`.

This matches `euclid::Rect`'s API, but it also seems a little better to
have strong typing without implicit conversion like this. In a few cases
it's not obvious that the argument is variable of tuple type, so there
is an implied assertion that the coordinate space is right.

The use of `#[deprecated]` makes this not entirely urgent for
compositors, but uses of this should be updated to `::new`, or other
functions like `from_size`.