wayland/viewporter: use current state for viewport checking
the viewport is checked as part of the commit handling at which
point the viewport has been moved to the current state.
using the pending state can lead to data races if the wayland
socket gets dispatched after receiving client updates but before
calling the commit handler. This can happen in a multi threaded
environment, but also when using blockers.

Regarding blockers:
Commit A does not contain an update to the viewport but is blocked.
Commit B contains an update to the viewport and sets the pending state but is delayed to after Commit B.
Commit A is unblocked and now uses the wrong viewport for validation.