rename Rectangle::from_extemities to Rectangle::from_extremeties
a function under the old name still exists, but it just calls
`from_extremeties` under the hood, is not on the documentation and is
deprecated, so that this is not a breaking change.

the actual word is called "extremities", extemities is just a
misspelling, see https://en.wiktionary.org/wiki/extremities.