text-input: fix active instance tracking
Only one text-input must be enabled at a time, by checking the
text-input id when issuing a request and comparing to the active one,
however the regular `Enabled` events must be send for all of them,
and as for the leave, we also send it for all the events to make it
symmetrical with the `enter.

Fixes: 475072d410df (text-input: Ensure only one enabled...)
Fixes #1604.