xwm: Update `override_redirect` in `MapNotify` handler
Apparently `XChangeWindowAttributes` can be used to change the
`override-redirect` flag for a window, and some clients create a window
without the flag, then set it this way. So we should handle the flag in
`MapNotify` to detect this.

Perhaps some other things should be changed, but this seems to help.
This fixes menus not showing up on cosmic-comp in REAPER and Intel
Questa, and presumably various other clients.