doc: Fix redundant-explicit-link errors
rustdoc::redundant-explicit-link matches more cases in the nightly rust
version. Remove the redundant links that trigger the warning, which is
processed as an error in CI due to `-D warnings` in the rustdoc flags.

The three redundant links all refer to items that are imported into the
current file and therefore do not require explicit, absolute links to
resolve.