render/multigpu: Don't use implicit modifier in `dma_shadow_copy`
I've been seeing `eglCreateImageKHR` errors here copying Nvidia buffers
to AMD. And worse, this somehow lead to a crash in the Nvidia driver.

It seems this was trying to use an implicit modifier, resulting in
something that won't import on the other GPU. Presumably we should just
filter out `Modifier::Invalid` here.