Add option to accept any node for direct scanout
This patch introduces a breaking change which allows restoring the
pre-0.7.0 functionality of acccepting any node for direct scanout for
the `GbmFramebufferExporter` constructor.

While it would be possible to do the same in a non-breaking manner by
introducing a separate constructor, this should be more elegant
long-term while requiring minimal breaking changes.