Fix the CI pipeline (#1769)
* Relax clippy warnings and unbreak ci

* clippy: Fix abs_diff warning

* doc: Fix redundant-explicit-link errors

rustdoc::redundant-explicit-link matches more cases in the nightly rust
version. Remove the redundant links that trigger the warning, which is
processed as an error in CI due to `-D warnings` in the rustdoc flags.

The three redundant links all refer to items that are imported into the
current file and therefore do not require explicit, absolute links to
resolve.

---------

Co-authored-by: dcz <gilapfco.dcz@porcupinefactory.org>
Co-authored-by: Tom Stokes <tomstokes@radixengineering.com>