multigpu: Don't re-use textures for re-created renderers
If a renderer gets re-created with the same api and same node,
existing cached `MultiTexture`s might be used with now outdated
and invalid texture ids.

This happens e.g. in cosmic-comp with destroys `EGLContext`s, when
a particular device isn't used anymore to give the driver a chance
to put the device into deeper sleep modes.

The obvious solution here is to not re-invent the wheel and only use
`TypeId`s + `DrmNode`s for hashing but our new `ErasedContextId`,
which will change for newly created renderers, that aren't based on
the same context.

This also simplifies the `MultiTextureInternal`-struct and related
code quite a bit, as we don't have a double nested `HashMap` anymore.