wayland/shell/xdg: Track last acked configure in MultiCache
The last-acked configure is part of the committed state, and should be tracked
as such. Before this commit, it would be possible to observe a newer "current
state" in a commit handler if the surface acked the next state while the
compositor was waiting on a blocker set in a pre-commit handler.

As part of the change, the current_state() getter was replaced by
with_committed_state(). Since the "current" state wasn't really a
correct thing most of the time, most existing uses of current_state()
actually wanted something different, and were corrected in previous
commits.