drm/output: Allow `DrmOutput` descruction during manager-lock
Since #1778 we provide an api to explicitly lock the `DrmOutputManager`
to apply multiple changes atomically without `DrmOutput`s being able to
observe any inconsistent state.

However as a side-effect of this, it has become impossible to re-assign
a crtc to a different connector in one atomic opertion, as the lock on
the `DrmOutputManager` prevents the `Drop`-implementation of `DrmOutput`
to free any crtcs for re-use.

This commit introduces a channel, if locking fails to deferr freeing
the crtc to the locked manager.