fix: Documentation of `Transform`
The current documentation for `Flipped` states that the "Plane is
flipped vertically", but the [wl_output::transform] specification
describes it as a "180 degree flip around a vertical axis". The
implementation follows the latter. Therefore, the documentation for
`Transform::Flipped` is incorrect and needs to be updated.

This patch corrects the documentation for `Transform` to remove
ambiguity and align with the implementation and the Wayland
specification.

For more details, see
https://github.com/kenoss/sabiniwm/blob/a4364ca211f97b672f92fb4de78f570f74404b90/note/issue-transform.md
and #Transform-Flipped-doc-wrong.

[wl_output::transform] https://wayland.freedesktop.org/docs/html/apa.html#protocol-spec-wl_output-enum-transform